mediflow/backend/mediflow_db/requirements.txt
# ============================================================
#  crimson/mediflow_db/requirements.txt
# ============================================================

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
